requirements.txt
cohere
langchain
langchain_cohere
streamlit
python-dotenv
PyPDF2

-e.